SpanLinks: extract trace-to-logs code to a module
