+++ public/app/features/explore/TraceView/createSpanLink.tsx
@@ -6,6 +6,7 @@ import {
   type DataSourceJsonData,
   dateTime,
   type Field,
+  FieldType,
   type LinkModel,
   mapInternalLinkToExplore,
   rangeUtil,
@@ -18,18 +19,17 @@ import {
   type TraceToProfilesOptions,
   type TraceToMetricsOptions,
   type TraceToLogsOptionsV2,
-  type TraceToLogsTag,
 } from '@grafana/o11y-ds-frontend';
 import { type PromQuery } from '@grafana/prometheus';
 import { getTemplateSrv } from '@grafana/runtime';
-import { type DataQuery } from '@grafana/schema';
 import { Icon } from '@grafana/ui';
 
-import { type LokiQuery } from '../../loki-helpers/types';
 import { type ExploreFieldLinkModel, getFieldLinksForExplore, getVariableUsageInfo } from '../utils/links';
 
+import { getFormattedTags, getSpanTags, getTraceToLogsQuery } from './components/logsLink';
 import { type SpanLinkDef, type SpanLinkFunc, SpanLinkType } from './components/types/links';
 import { type Trace, type TraceSpan, type TraceSpanReference } from './components/types/trace';
+import { getDefaultMetricTags, getDefaultProfilingTags } from './crossSignalConfig';
 
 /**
  * This is a factory for the link creator. It returns the function mainly so it can return undefined in which case
@@ -141,14 +141,8 @@ export function createSpanLinkFactory({
 /**
  * Default keys to use when there are no configured tags.
  */
-const formatDefaultKeys = (keys: string[]) => {
-  return keys.map((k) => ({
-    key: k,
-    value: k.includes('.') ? k.replace('.', '_') : undefined,
-  }));
-};
-const defaultKeys = formatDefaultKeys(['cluster', 'hostname', 'namespace', 'pod', 'service.name', 'service.namespace']);
-export const defaultProfilingKeys = formatDefaultKeys(['service.name', 'service.namespace']);
+
+export const defaultProfilingKeys = getDefaultProfilingTags();
 export const pyroscopeProfileIdTagKey = 'pyroscope.profile.id';
 const feO11yTagKey = 'gf.feo11y.app.id';
 
@@ -164,56 +158,22 @@ function legacyCreateSpanLinkFactory(
   logsDataSourceSettings?: DataSourceInstanceSettings<DataSourceJsonData>,
   metricsDataSourceSettings?: DataSourceInstanceSettings<DataSourceJsonData>
 ) {
-  const isSplunkDS = logsDataSourceSettings?.type === 'grafana-splunk-datasource';
-
   return function SpanLink(span: TraceSpan): SpanLinkDef[] {
     scopedVars = {
       ...scopedVars,
       ...scopedVarsFromSpan(span),
     };
     const links: SpanLinkDef[] = [];
-    let query: DataQuery | undefined;
-    let tags = '';
 
     // TODO: This should eventually move into specific data sources and added to the data frame as we no longer use the
     //  deprecated blob format and we can map the link easily in data frame.
     if (logsDataSourceSettings && traceToLogsOptions) {
-      const customQuery = traceToLogsOptions.customQuery ? traceToLogsOptions.query : undefined;
-      const tagsToUse =
-        traceToLogsOptions.tags && traceToLogsOptions.tags.length > 0 ? traceToLogsOptions.tags : defaultKeys;
-      switch (logsDataSourceSettings?.type) {
-        case 'loki':
-          tags = getFormattedTags(span, tagsToUse);
-          query = getQueryForLoki(span, traceToLogsOptions, tags, customQuery);
-          break;
-        case 'grafana-splunk-datasource':
-          tags = getFormattedTags(span, tagsToUse, { joinBy: ' ' });
-          query = getQueryForSplunk(span, traceToLogsOptions, tags, customQuery);
-          break;
-        case 'elasticsearch':
-        case 'grafana-opensearch-datasource':
-          tags = getFormattedTags(span, tagsToUse, { labelValueSign: ':', joinBy: ' AND ' });
-          query = getQueryForElasticsearchOrOpensearch(span, traceToLogsOptions, tags, customQuery);
-          break;
-        case 'grafana-falconlogscale-datasource':
-          tags = getFormattedTags(span, tagsToUse, { joinBy: ' OR ' });
-          query = getQueryForFalconLogScale(span, traceToLogsOptions, tags, customQuery);
-          break;
-        case 'googlecloud-logging-datasource':
-          tags = getFormattedTags(span, tagsToUse, { joinBy: ' AND ' });
-          query = getQueryForGoogleCloudLogging(span, traceToLogsOptions, tags, customQuery);
-          break;
-        case 'victoriametrics-logs-datasource':
-          // Build tag selector using strict equality (":=") required by LogsQL
-          // See https://docs.victoriametrics.com/victorialogs/logsql/#exact-filter
-          tags = getFormattedTags(span, tagsToUse, { labelValueSign: ':=', joinBy: ' AND ' });
-          query = getQueryForVictoriaLogs(span, traceToLogsOptions, tags, customQuery);
-          break;
-      }
+      const { query, tags } = getTraceToLogsQuery(logsDataSourceSettings, traceToLogsOptions);
 
       // query can be false in case the simple UI tag mapping is used but none of them are present in the span.
       // For custom query, this is always defined and we check if the interpolation matched all variables later on.
       if (query) {
+        const isSplunkDS = logsDataSourceSettings?.type === 'grafana-splunk-datasource';
         const dataLink: DataLink = {
           title: logsDataSourceSettings.name,
           url: '',
@@ -252,7 +212,12 @@ function legacyCreateSpanLinkFactory(
             internalLink: dataLink.internal!,
             scopedVars: scopedVars,
             range: dataLink.internal!.range,
-            field: {} as Field,
+            field: {
+              name: '',
+              type: FieldType.other,
+              config: {},
+              values: [],
+            },
             onClickFn: splitOpenFn,
             replaceVariables: getTemplateSrv().replace.bind(getTemplateSrv()),
           });
@@ -313,13 +278,13 @@ function legacyCreateSpanLinkFactory(
         const tagsToUse =
           traceToMetricsOptions.tags && traceToMetricsOptions.tags.length > 0
             ? traceToMetricsOptions.tags
-            : defaultKeys;
+            : getDefaultMetricTags();
 
         scopedVars = {
           ...scopedVars,
           __tags: {
             text: t('explore.legacy-create-span-link-factory.text.tags', 'Tags'),
-            value: getFormattedTags(span, tagsToUse),
+            value: getFormattedTags(getSpanTags(span), tagsToUse),
           },
         };
 
@@ -335,7 +300,12 @@ function legacyCreateSpanLinkFactory(
               ? rangeUtil.intervalToMs(traceToMetricsOptions.spanEndTimeShift)
               : 120000,
           }),
-          field: {} as Field,
+          field: {
+            name: '',
+            type: FieldType.other,
+            config: {},
+            values: [],
+          },
           onClickFn: splitOpenFn,
           replaceVariables: getTemplateSrv().replace.bind(getTemplateSrv()),
         });
@@ -427,38 +397,6 @@ const getReferenceTitle = (reference: TraceSpanReference) => {
   return title;
 };
 
-function getQueryForLoki(
-  span: TraceSpan,
-  options: TraceToLogsOptionsV2,
-  tags: string,
-  customQuery?: string
-): LokiQuery | undefined {
-  const { filterByTraceID, filterBySpanID } = options;
-
-  if (customQuery) {
-    return { expr: customQuery, refId: '' };
-  }
-
-  if (!tags) {
-    return undefined;
-  }
-
-  let expr = '{${__tags}}';
-  if (filterByTraceID && span.traceID) {
-    expr +=
-      ' | label_format log_line_contains_trace_id=`{{ contains "${__span.traceId}" __line__  }}` | log_line_contains_trace_id="true" or trace_id="${__span.traceId}"';
-  }
-  if (filterBySpanID && span.spanID) {
-    expr +=
-      ' | label_format log_line_contains_span_id=`{{ contains "${__span.spanId}" __line__  }}` | log_line_contains_span_id="true" or span_id="${__span.spanId}"';
-  }
-
-  return {
-    expr: expr,
-    refId: '',
-  };
-}
-
 function getLinkForFeO11y(span: TraceSpan): string | undefined {
   const feO11yAppId = span.process.tags.find((tag) => tag.key === feO11yTagKey)?.value;
   const feO11ySessionId = span.tags.find((tag) => tag.key === 'session_id' || tag.key === 'session.id')?.value;
@@ -466,204 +404,6 @@ function getLinkForFeO11y(span: TraceSpan): string | undefined {
   return feO11yAppId && feO11ySessionId
     ? `/a/grafana-kowalski-app/apps/${feO11yAppId}/sessions/${feO11ySessionId}`
     : undefined;
-}
-
-// we do not have access to the dataquery type for opensearch,
-// so here is a minimal interface that handles both elasticsearch and opensearch.
-interface ElasticsearchOrOpensearchQuery extends DataQuery {
-  query: string;
-  metrics: Array<{
-    id: string;
-    type: 'logs';
-  }>;
-}
-
-function getQueryForElasticsearchOrOpensearch(
-  span: TraceSpan,
-  options: TraceToLogsOptionsV2,
-  tags: string,
-  customQuery?: string
-): ElasticsearchOrOpensearchQuery {
-  const { filterByTraceID, filterBySpanID } = options;
-  if (customQuery) {
-    return {
-      query: customQuery,
-      refId: '',
-      metrics: [{ id: '1', type: 'logs' }],
-    };
-  }
-
-  let queryArr = [];
-  if (filterBySpanID && span.spanID) {
-    queryArr.push('"${__span.spanId}"');
-  }
-
-  if (filterByTraceID && span.traceID) {
-    queryArr.push('"${__span.traceId}"');
-  }
-
-  if (tags) {
-    queryArr.push('${__tags}');
-  }
-
-  return {
-    query: queryArr.join(' AND '),
-    refId: '',
-    metrics: [{ id: '1', type: 'logs' }],
-  };
-}
-
-function getQueryForSplunk(span: TraceSpan, options: TraceToLogsOptionsV2, tags: string, customQuery?: string) {
-  const { filterByTraceID, filterBySpanID } = options;
-
-  if (customQuery) {
-    return { query: customQuery, refId: '' };
-  }
-
-  let query = '';
-  if (tags) {
-    query += '${__tags}';
-  }
-  if (filterByTraceID && span.traceID) {
-    query += ' "${__span.traceId}"';
-  }
-  if (filterBySpanID && span.spanID) {
-    query += ' "${__span.spanId}"';
-  }
-
-  return {
-    query: query,
-    refId: '',
-  };
-}
-
-function getQueryForGoogleCloudLogging(
-  span: TraceSpan,
-  options: TraceToLogsOptionsV2,
-  tags: string,
-  customQuery?: string
-) {
-  const { filterByTraceID, filterBySpanID } = options;
-
-  if (customQuery) {
-    return { query: customQuery, refId: '' };
-  }
-
-  let queryArr = [];
-  if (filterBySpanID && span.spanID) {
-    queryArr.push('"${__span.spanId}"');
-  }
-
-  if (filterByTraceID && span.traceID) {
-    queryArr.push('"${__span.traceId}"');
-  }
-
-  if (tags) {
-    queryArr.push('${__tags}');
-  }
-
-  return {
-    query: queryArr.join(' AND '),
-    refId: '',
-  };
-}
-
-function getQueryForFalconLogScale(span: TraceSpan, options: TraceToLogsOptionsV2, tags: string, customQuery?: string) {
-  const { filterByTraceID, filterBySpanID } = options;
-
-  if (customQuery) {
-    return {
-      lsql: customQuery,
-      refId: '',
-    };
-  }
-
-  if (!tags) {
-    return undefined;
-  }
-
-  let lsql = '${__tags}';
-  if (filterByTraceID && span.traceID) {
-    lsql += ' or "${__span.traceId}"';
-  }
-
-  if (filterBySpanID && span.spanID) {
-    lsql += ' or "${__span.spanId}"';
-  }
-
-  return {
-    lsql,
-    refId: '',
-  };
-}
-
-/**
- * Builds a LogsQL expression for victoria‑metrics‑logs‑datasource.
- * Uses := for exact‑match filters and joins parts with AND.
- * See https://docs.victoriametrics.com/victorialogs/logsql/#exact-filter
- */
-function getQueryForVictoriaLogs(span: TraceSpan, options: TraceToLogsOptionsV2, tags: string, customQuery?: string) {
-  const { filterByTraceID, filterBySpanID } = options;
-
-  // Custom user query has priority
-  if (customQuery) {
-    return {
-      expr: customQuery,
-      refId: '',
-    };
-  }
-
-  const parts: string[] = [];
-
-  if (filterBySpanID && span.spanID) {
-    parts.push('span_id:="${__span.spanId}"');
-  }
-  if (filterByTraceID && span.traceID) {
-    parts.push('trace_id:="${__span.traceId}"');
-  }
-  if (tags) {
-    parts.push('${__tags}');
-  }
-
-  // Nothing to match against – do not create the link
-  if (!parts.length) {
-    return undefined;
-  }
-
-  return {
-    expr: parts.join(' AND '),
-    refId: '',
-  };
-}
-
-/**
- * Creates a string representing all the tags already formatted for use in the query. The tags are filtered so that
- * only intersection of tags that exist in a span and tags that you want are serialized into the string.
- */
-export function getFormattedTags(
-  span: TraceSpan,
-  tags: TraceToLogsTag[],
-  { labelValueSign = '=', joinBy = ', ' }: { labelValueSign?: string; joinBy?: string } = {}
-) {
-  // In order, try to use mapped tags -> tags -> default tags
-  // Build tag portion of query
-  return [
-    ...span.process.tags,
-    ...span.tags,
-    { key: 'spanId', value: span.spanID },
-    { key: 'traceId', value: span.traceID },
-    { key: 'name', value: span.operationName },
-    { key: 'duration', value: span.duration },
-  ]
-    .map((tag) => {
-      const keyValue = tags.find((keyValue) => keyValue.key === tag.key);
-      if (keyValue) {
-        return `${keyValue.value ? keyValue.value : keyValue.key}${labelValueSign}"${tag.value}"`;
-      }
-      return undefined;
-    })
-    .filter((v) => Boolean(v))
-    .join(joinBy);
 }
 
 /**
@@ -771,7 +511,7 @@ export function scopedVarsFromTags(
     tags = {
       __tags: {
         text: t('explore.scoped-vars-from-tags.text.tags', 'Tags'),
-        value: getFormattedTags(span, profileTags),
+        value: getFormattedTags(getSpanTags(span), profileTags),
       },
     };
   }
